Fix `test_invalid_utf8_filename` test on Windows.
On Windows, Python gives me the following error:

UnicodeDecodeError: 'utf-8' codec can't decode byte 0xc3 in position 0:
 invalid continuation byte

for the `open` call which creates a file with an invalid Unicode byte
string. This works fine on Linux but does not work on Windows.
Looks like the call first checks if the byte string is valid Unicode
or not.

The actual filtering in `lib50\_api.py` is unchanged (for now).