Turn off input echo for passwords on Windows.
This is one of the major issues in porting lib50 to Windows, along with
subprocess output management.

The workaround is sourced from https://stackoverflow.com/a/9218551 .
This workaround uses the Python `ctypes` library to call the Windows
library functions `SetConsoleMode` and `GetConsoleMode` functions
with the modes specified in the documentation at
https://learn.microsoft.com/en-us/windows/console/setconsolemode .

Code takes the current console mode and stores it. Then it turns off the
ENABLE_ECHO_INPUT flag(these values are *hardcoded*) and yields.
The console mode is set to the old value after the context is exited.