`PopenSpawn` and `psutil` replace `pexpect.spawn`.
`pexpect` works with Unix `pty` pseudo-terminals and hence not available
on Windows.

The functionality we use is
- `expect` for output matching
- accessing exit code with `exitstatus` attribute
- `isalive` for child process status

`PopenSpawn` uses `subprocess.spawn` which is cross platform.
`proc` of `PopenSpawn` returns the Python's native `Proc` object that
has attributes for exit code, child process status, waiting functions
and the `pexpect` library implements the output matching.