Fix Git Credential caching on Windows.
Git on Windows does not support Unix sockets(it is not compiled with
them on Windows 10 and previous Windows versions do not have it) which
makes the `credential.helper=cache` invalid.

Switched to Git Credential Manager for Windows.

Reference: https://stackoverflow.com/a/5343146

An additional check for output of GCM Core is needed as lines endings
are with `\n` and not with `\r\n`.